requirements.txt
pandas
numpy
streamlit
joblib
scikit-learn
matplotlib
seaborn
plotly
dvc 
dvc-s3
PyYAML
